requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
langchain-text-splitters
langchain-core
sentence-transformers
chromadb
pdfplumber
pymupdf
python-dotenv